[SLAC22-227] Move config into module
Remove config from the sync directory so we only install the migration
configurations that we need. Set a module dependency for the configs in
the install folder so that the module can be uninstalled and re-installed
to update the config in the sync directory. This also adds a core hash to
the config files in the sync directory to track changes from the module.